小时定时任务增加 =更新日耗电总量

+++ src/main/java/com/example/springboot3_2vue3/mapper/equipmapper/YearpoweruseMapper.java
@@ -9,5 +9,5 @@ import com.example.springboot3_2vue3.domain.equipment.Yearpoweruse;
  * @DateTime: 2023/6/17 10:56
  **/
 public interface YearpoweruseMapper extends BaseMapper<Yearpoweruse> {
-   int InsertOrUpdatePoweruse(int poweruse);
+   int InsertOrUpdatePoweruse(float poweruse);
 }
